move generate message id to own file
